pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gic-plan-to-perform"
version = "0.2.0"
description = "AI-Powered Plan-to-Perform: Commodity Forecast, Financial Driver Engine & Real-Time Market Intelligence"
requires-python = ">=3.10"
dependencies = [
    # ── Data Processing ──
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # ── ML & Statistics ──
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "xgboost>=2.0.0",

    # ── Financial Data Sources ──
    "yfinance>=0.2.36",
    "ccxt>=4.2.0",
    "ffn>=1.0.0",
    "fredapi>=0.5.1",

    # ── API & Web ──
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",

    # ── Dashboard ──
    "streamlit>=1.30.0",

    # ── Visualization ──
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",

    # ── Infrastructure ──
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "joblib>=1.3.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
notebook = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
